+++ weldx/tests/transformations/test_cs_manager.py
@@ -12,14 +12,6 @@ import xarray as xr
 from pandas import TimedeltaIndex as TDI  # noqa
 from pandas import Timestamp as TS  # noqa
 from pandas import date_range
-from ._test_util import (
-    r_mat_x,
-    r_mat_y,
-    r_mat_z,
-    check_coordinate_system_orientation,
-    check_coordinate_system,
-    check_cs_close,
-)
 
 import weldx.transformations as tf
 import weldx.util as ut
@@ -29,6 +21,15 @@ from weldx.tests._helpers import get_test_name
 from weldx.time import Time, types_time_like, types_timestamp_like
 from weldx.transformations import LocalCoordinateSystem as LCS  # noqa
 from weldx.transformations import WXRotation
+
+from ._util import (
+    check_coordinate_system,
+    check_coordinate_system_orientation,
+    check_cs_close,
+    r_mat_x,
+    r_mat_y,
+    r_mat_z,
+)
 
 CSM = tf.CoordinateSystemManager
 LCS = tf.LocalCoordinateSystem
