+++ weldx/tests/transformations/test_local_cs.py
@@ -5,14 +5,6 @@ import math
 import random
 from copy import deepcopy
 from typing import Any, Dict, List, Union
-from ._test_util import (
-    r_mat_x,
-    r_mat_y,
-    r_mat_z,
-    check_coordinate_system,
-    check_coordinate_system_orientation,
-    check_cs_close,
-)
 
 import numpy as np
 import pandas as pd
@@ -30,6 +22,15 @@ from weldx.tests._helpers import get_test_name
 from weldx.time import Time, types_time_like, types_timestamp_like
 from weldx.transformations import LocalCoordinateSystem as LCS  # noqa
 from weldx.transformations import WXRotation
+
+from ._util import (
+    check_coordinate_system,
+    check_coordinate_system_orientation,
+    check_cs_close,
+    r_mat_x,
+    r_mat_y,
+    r_mat_z,
+)
 
 
 class TestLocalCoordinateSystem:
